benchmark/keras/requirements.txt
tensorflow-macos==2.10.*
tensorflow-metal==0.6.*
pandas==1.4.*
scikit-learn==1.1.*
matplotlib==3.5.*